backend/requirements.txt
fastapi==0.65.1
uvicorn==0.14.0
pymongo==3.11.4
jira==3.0.1
python-dotenv==0.17.1